+++ R/apply_fmt_fn.R
@@ -11,19 +11,19 @@
 #'
 #' @examples
 #' ard_continuous(ADSL, variables = "AGE") |>
-#'   apply_statistic_fmt_fn()
-apply_statistic_fmt_fn <- function(x) {
+#'   apply_fmt_fn()
+apply_fmt_fn <- function(x) {
   if (!inherits(x, "card")) {
     cli::cli_abort(c("i" = "Argument {.code x} must be class {.cls card}."))
   }
 
   x |>
     dplyr::mutate(
-      .after = "statistic",
-      statistic_fmt =
+      .after = "stat",
+      stat_fmt =
         map2(
-          .data$statistic,
-          .data$statistic_fmt_fn,
+          .data$stat,
+          .data$fmt_fn,
           function(x, fn) {
             if (!is.null(fn)) {
               do.call(alias_as_fmt_fn(fn), args = list(x))