+++ R/ard_categorical.R
@@ -24,8 +24,8 @@
 #'   a named list, a list of formulas,
 #'   or a single formula where the list element is a named list of functions
 #'   (or the RHS of a formula),
-#'   e.g. `list(mpg = categorical_variable_summary_fns())`.
-#' @param stat_labels ([`formula-list-selector`][syntax])\cr
+#'   e.g. `list(mpg = categorical_summary_fns())`.
+#' @param stat_label ([`formula-list-selector`][syntax])\cr
 #'   a named list, a list of formulas, or a single formula where
 #'   the list element is either a named list or a list of formulas defining the
 #'   statistic labels, e.g. `everything() ~ list(n = "n", p = "pct")` or
@@ -68,16 +68,16 @@
 #'   dplyr::group_by(ARM) |>
 #'   ard_categorical(
 #'     variables = "AGEGR1",
-#'     statistic = everything() ~ categorical_variable_summary_fns("n")
+#'     statistic = everything() ~ categorical_summary_fns("n")
 #'   )
 ard_categorical <- function(data,
                             variables,
                             by = dplyr::group_vars(data),
                             strata = NULL,
-                            statistic = everything() ~ categorical_variable_summary_fns(),
+                            statistic = everything() ~ categorical_summary_fns(),
                             denominator = NULL,
                             fmt_fn = NULL,
-                            stat_labels = everything() ~ default_stat_labels()) {
+                            stat_label = everything() ~ default_stat_labels()) {
   # check inputs ---------------------------------------------------------------
   check_not_missing(data)
   check_not_missing(variables)
@@ -97,7 +97,7 @@ ard_categorical <- function(data,
   process_formula_selectors(
     data = data[variables],
     statistic = statistic,
-    stat_labels = stat_labels,
+    stat_label = stat_label,
     fmt_fn = fmt_fn
   )
   fill_formula_selectors(
@@ -127,9 +127,9 @@ ard_categorical <- function(data,
         strata = all_of(strata),
         statistic = statistics_non_tabulation,
         fmt_fn = NULL,
-        stat_labels = NULL
+        stat_label = NULL
       ) |>
-      dplyr::select(-c("stat_label", "statistic_fmt_fn"))
+      dplyr::select(-c("stat_label", "fmt_fn"))
   }
 
   # calculate tabulation statistics
@@ -149,7 +149,7 @@ ard_categorical <- function(data,
     dplyr::bind_rows(df_result_tabulation, df_result_non_tabulation) |>
     .process_nested_list_as_df(
       arg = fmt_fn,
-      new_column = "statistic_fmt_fn"
+      new_column = "fmt_fn"
     ) |>
     .default_fmt_fn()
 
@@ -157,7 +157,7 @@ ard_categorical <- function(data,
   df_result_final <-
     .process_nested_list_as_df(
       x = df_result_final,
-      arg = stat_labels,
+      arg = stat_label,
       new_column = "stat_label",
       unlist = TRUE
     ) |>
@@ -266,7 +266,7 @@ ard_categorical <- function(data,
           tidyr::pivot_longer(
             cols = any_of(c("...ard_n...", "...ard_N...", "...ard_p...")),
             names_to = "stat_name",
-            values_to = "statistic"
+            values_to = "stat"
           ) |>
           dplyr::mutate(
             stat_name =