+++ R/ard_dichotomous.R
@@ -18,17 +18,17 @@
 #'   ard_dichotomous(
 #'     variables = c(cyl, am),
 #'     value = list(cyl = 4),
-#'     statistic = ~ categorical_variable_summary_fns("p")
+#'     statistic = ~ categorical_summary_fns("p")
 #'   )
 ard_dichotomous <- function(data,
                             variables,
                             by = dplyr::group_vars(data),
                             strata = NULL,
                             value = maximum_variable_value(data[variables]),
-                            statistic = everything() ~ categorical_variable_summary_fns(),
+                            statistic = everything() ~ categorical_summary_fns(),
                             denominator = NULL,
                             fmt_fn = NULL,
-                            stat_labels = everything() ~ default_stat_labels()) {
+                            stat_label = everything() ~ default_stat_labels()) {
   # check inputs ---------------------------------------------------------------
   check_not_missing(data)
   check_not_missing(variables)
@@ -57,7 +57,7 @@ ard_dichotomous <- function(data,
     statistic = statistic,
     denominator = denominator,
     fmt_fn = fmt_fn,
-    stat_labels = stat_labels
+    stat_label = stat_label
   ) |>
     dplyr::filter(
       pmap(