+++ R/tidy_ard_order.R
@@ -39,7 +39,7 @@ tidy_ard_column_order <- function(x) {
     all_ard_variables(),
     any_of(c(
       "context",
-      "stat_name", "stat_label", "statistic", "statistic_fmt", "statistic_fmt_fn",
+      "stat_name", "stat_label", "stat", "stat_fmt", "fmt_fn",
       "warning", "error"
     )),
     dplyr::everything()