+++ tests/testthat/test-shuffle_ard.R
@@ -48,20 +48,20 @@ test_that("shuffle/trim works", {
     as.data.frame()
   expect_snapshot(ard_shuff_trim)
   # only numeric stats
-  expect_type(ard_shuff_trim$statistic, "double")
+  expect_type(ard_shuff_trim$stat, "double")
   # no list columns
   expect_true(!any(map_lgl(ard_shuff_trim, is.list)))
 })
 
 test_that("shuffle_ard handles protected names", {
   ard_test <- ard_categorical(
-    ADSL |> dplyr::rename(statistic = ARM),
-    by = "statistic",
+    ADSL |> dplyr::rename(stat = ARM),
+    by = "stat",
     variables = "AGEGR1"
   ) |>
     shuffle_ard()
 
-  expect_equal(names(ard_test)[1], "statistic.1")
+  expect_equal(names(ard_test)[1], "stat.1")
 })
 
 test_that("shuffle_ard notifies user about warnings/errors before dropping", {
@@ -78,8 +78,8 @@ test_that("shuffle_ard fills missing group levels if the group is meaningful", {
   # mix of missing/nonmissing group levels present before shuffle
   expect_snapshot(
     bind_ard(
-      ard_continuous(ADSL, by = "ARM", variables = "AGE", statistic = ~ continuous_variable_summary_fns("mean")),
-      dplyr::tibble(group1 = "ARM", variable = "AGE", stat_name = "p", stat_label = "p", statistic = list(0.05))
+      ard_continuous(ADSL, by = "ARM", variables = "AGE", statistic = ~ continuous_summary_fns("mean")),
+      dplyr::tibble(group1 = "ARM", variable = "AGE", stat_name = "p", stat_label = "p", stat = list(0.05))
     ) |>
       shuffle_ard()
   )
@@ -87,8 +87,8 @@ test_that("shuffle_ard fills missing group levels if the group is meaningful", {
   # no group levels present before shuffle
   expect_snapshot(
     bind_ard(
-      ard_continuous(ADSL, variables = "AGE", statistic = ~ continuous_variable_summary_fns("mean")),
-      dplyr::tibble(group1 = "ARM", variable = "AGE", stat_name = "p", stat_label = "p", statistic = list(0.05))
+      ard_continuous(ADSL, variables = "AGE", statistic = ~ continuous_summary_fns("mean")),
+      dplyr::tibble(group1 = "ARM", variable = "AGE", stat_name = "p", stat_label = "p", stat = list(0.05))
     ) |>
       shuffle_ard()
   )
@@ -100,10 +100,10 @@ test_that("shuffle_ard doesn't trim off NULL/NA values", {
     data.frame(x = rep_len(NA_character_, 10)) |>
       ard_continuous(
         variables = x,
-        statistic = ~ continuous_variable_summary_fns(c("median", "p25", "p75"))
+        statistic = ~ continuous_summary_fns(c("median", "p25", "p75"))
       ) |>
       shuffle_ard() |>
-      dplyr::pull(statistic)
+      dplyr::pull(stat)
   )
 
   # check that all rows present and result is a numeric vector